v2 ssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssss

+++ ft_memcmp.c
@@ -1,8 +1,20 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_memcmp.c                                        :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: chevrethis <chevrethis@student.42.fr>      +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/04/29 13:14:52 by chevrethis        #+#    #+#             */
+/*   Updated: 2025/04/29 13:17:37 by chevrethis       ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include "libft.h"
 
 void    *ft_memcmp(const void *s1, const void *s2, size_t n)
 {
-    int i;
+    size_t i;
 
     i = 0;
     while ((*(unsigned char*)(s1 + i) != '\0' || *(unsigned char*)(s2 + i) != '\0') && i < n)
